clear Main class
